fix: [Simple Agent] In simple agents, workflow work is used as a skill, and no specified skill is called during dialogue
