+++ apps/application/flow/step_node/ai_chat_step_node/impl/base_chat_node.py
@@ -9,123 +9,25 @@
 import json
 import re
 import time
-import uuid
 from functools import reduce
 from typing import List, Dict
 
-import uuid_utils.compat as uuid
-from django.db.models import QuerySet, OuterRef, Subquery
+from django.db.models import QuerySet
 from django.utils.translation import gettext as _
 from langchain_core.messages import BaseMessage, AIMessage, HumanMessage, SystemMessage
-from langchain_core.tools import StructuredTool
-from pydantic import Field, create_model
 
-from application.flow.common import Workflow, WorkflowMode
-from application.flow.i_step_node import NodeResult, INode, ToolWorkflowPostHandler
+from application.flow.common import WorkflowMode
+from application.flow.i_step_node import NodeResult, INode
 from application.flow.step_node.ai_chat_step_node.i_chat_node import IChatNode
-from application.flow.tools import Reasoning, mcp_response_generator
+from application.flow.tools import Reasoning, mcp_response_generator, get_tools
 from application.models import Application, ApplicationApiKey, ApplicationAccessToken
-from application.serializers.common import ToolExecute
 from common.exception.app_exception import AppApiException
 from common.utils.rsa_util import rsa_long_decrypt
 from common.utils.shared_resource_auth import filter_authorized_ids
 from common.utils.tool_code import ToolExecutor
 from models_provider.models import Model
 from models_provider.tools import get_model_credential, get_model_instance_by_model_workspace_id
-from tools.models import Tool, ToolWorkflowVersion, ToolType
-
-
-def build_schema(fields: dict):
-    return create_model("dynamicSchema", **fields)
-
-
-def get_type(_type: str):
-    if _type == 'float':
-        return float
-    if _type == 'string':
-        return str
-    if _type == 'int':
-        return int
-    if _type == 'dict':
-        return dict
-    if _type == 'array':
-        return list
-    if _type == 'boolean':
-        return bool
-    return object
-
-
-def get_workflow_args(tool, qv):
-    for node in qv.work_flow.get('nodes'):
-        if node.get('type') == 'tool-base-node':
-            input_field_list = node.get('properties').get('user_input_field_list')
-            return build_schema(
-                {field.get('field'): (get_type(field.get('type')), Field(..., description=field.get('desc')))
-                 for field in input_field_list})
-
-    return build_schema({})
-
-
-def get_workflow_func(node, tool, qv, workspace_id):
-    tool_id = tool.id
-    tool_record_id = str(uuid.uuid7())
-    took_execute = ToolExecute(tool_id, tool_record_id,
-                               workspace_id,
-                               node.workflow_manage.get_source_type(),
-                               node.workflow_manage.get_source_id(),
-                               False)
-
-    def inner(**kwargs):
-        from application.flow.tool_workflow_manage import ToolWorkflowManage
-        work_flow_manage = ToolWorkflowManage(
-            Workflow.new_instance(qv.work_flow, WorkflowMode.TOOL),
-            {
-                'chat_record_id': tool_record_id,
-                'tool_id': tool_id,
-                'stream': True,
-                'workspace_id': workspace_id,
-                **kwargs},
-
-            ToolWorkflowPostHandler(took_execute, tool_id),
-            is_the_task_interrupted=lambda: False,
-            child_node=None,
-            start_node_id=None,
-            start_node_data=None,
-            chat_record=None
-        )
-        res = work_flow_manage.run()
-        for r in res:
-            pass
-        return work_flow_manage.out_context
-
-    return inner
-
-
-def get_tools(node, tool_workflow_ids, workspace_id):
-    tools = QuerySet(Tool).filter(id__in=tool_workflow_ids, tool_type=ToolType.WORKFLOW, workspace_id=workspace_id)
-    latest_subquery = ToolWorkflowVersion.objects.filter(
-        tool_id=OuterRef('tool_id')
-    ).order_by('-create_time')
-
-    qs = ToolWorkflowVersion.objects.filter(
-        tool_id__in=[t.id for t in tools],
-        id=Subquery(latest_subquery.values('id')[:1])
-    )
-    qd = {q.tool_id: q for q in qs}
-    results = []
-    for tool in tools:
-        qv = qd.get(tool.id)
-        func = get_workflow_func(node, tool, qv, workspace_id)
-        args = get_workflow_args(tool, qv)
-        tool = StructuredTool.from_function(
-            func=func,
-            name=tool.name,
-            description=tool.desc,
-            args_schema=args,
-        )
-        results.append(tool)
-
-    return results
+from tools.models import Tool, ToolType
 
 
 def _write_context(node_variable: Dict, workflow_variable: Dict, node: INode, workflow, answer: str,
@@ -362,7 +264,8 @@ class BaseChatNode(IChatNode):
                     mcp_servers_config = {**mcp_servers_config, **json.loads(mcp_tool['code'])}
                     mcp_servers_config = self.handle_variables(mcp_servers_config)
         tool_init_params = {}
-        tools = get_tools(self, tool_ids, workspace_id)
+        tools = get_tools(self.workflow_manage.get_source_type(), self.workflow_manage.get_source_id(), tool_ids,
+                          workspace_id)
         if tool_ids and len(tool_ids) > 0:  # 如果有工具ID，则将其转换为MCP
             self.context['tool_ids'] = tool_ids
             for tool_id in tool_ids:
